requirements.txt
torch>=1.12
torchvision
torchaudio
transformers
scikit-learn
pandas
tqdm
pyyaml
matplotlib